src/requirements.txt
fastapi
uvicorn
requests
pyyaml
pytz
cryptography
